sort app.ts moved handleTermSignals to utils.ts
